requirements.txt
streamlit
psycopg2-binary
bcrypt
requests
scikit-surprise
streamlit-authenticator
